backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pycryptodome==3.19.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
langchain==0.0.350
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.0
vectorize-client>=0.2.1 